Create base Searcher classes-OligoSearcher, OligoSearcherMinimizeNumber, OligoSearcherMaximizeActivity, and GuideSearcher-to make GuideSearcherMinimizeGuides, GuideSearcherMaximizeActivity, and PrimerSearcher. Simplify some of the code underlying these classes Change guide_is_suitable_fn (previously used for specificity check) to is_suitable_fns-a list of suitability functions-and make the check_gc_content a part of this list Rename guide.py in utils to oligo.py-adjust design.py, alignment.py, coverage_analysis.py, alignment_query.py, target_search.py, pick_test_targets.py accordingly Move construct_guide from alignment.py to Guide/Primer searchers
**TESTS DO NOT WORK YET** Documentation is a work in progress
